refactor: insert 可能なlinksに変換する箇所を関数に変更
